fixes preview panel by adding additional checks
