+++ packages/notifi-react-card/lib/hooks/useFetchedCardState.ts
@@ -1,6 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
-
-import { useNotifiSubscriptionContext } from '../context';
+import { useState } from 'react';
 
 export type PreviewView = Readonly<{
   state: 'preview';
@@ -15,34 +13,12 @@ export type EditInfoView = Readonly<{
 export type FetchedCardView = PreviewView | HistoryView | EditInfoView;
 
 export const useFetchedCardState = () => {
-  const { email, phoneNumber, telegramId } = useNotifiSubscriptionContext();
-
   const [cardView, setCardView] = useState<FetchedCardView>({
     state: 'edit',
   });
 
-  const firstLoad = useRef(false);
-
-  useEffect(() => {
-    if (firstLoad.current) {
-      return;
-    }
-
-    firstLoad.current = true;
-
-    // could be undefined for FTU, so we need to check for undefined as well
-    if (
-      (email !== '' && email !== undefined) ||
-      (phoneNumber !== '' && phoneNumber !== undefined) ||
-      (telegramId !== '' && telegramId !== undefined)
-    ) {
-      setCardView({ state: 'preview' });
-    }
-  }, [email, phoneNumber, telegramId, setCardView, cardView]);
-
   return {
     cardView,
     setCardView,
-    useNotifiSubscriptionContext,
   };
 };
